print last digit

+++ 0x02-functions_nested_loops/main.h
@@ -21,7 +21,7 @@ int print_sign(int n);
 
 int _abs(int n);
 
-int print_last_digit(int n);
+int print_last_digit(long int n);
 
 void print_minutes(int i, int j, int k, int l);
